Add randomly_pick_reviewers function to make the main function slimmer
