+++ src/reviewer.js
@@ -2,6 +2,7 @@
 
 const core = require('@actions/core');
 const minimatch = require('minimatch');
+const sampleSize = require('lodash/sampleSize');
 
 function fetch_other_group_members({ author, config }) {
   const DEFAULT_OPTIONS = {
@@ -110,6 +111,23 @@ function fetch_default_reviwers({ config, excludes = [] }) {
   return [ ...new Set(individuals) ].filter((reviewer) => !excludes.includes(reviewer));
 }
 
+function randomly_pick_reviewers({ reviewers, config }) {
+  const DEFAULT_OPTIONS = {
+    numberOfReviewers: 0,
+  };
+
+  const { numberOfReviewers } = {
+    ...DEFAULT_OPTIONS,
+    ...config.options,
+  };
+
+  if (numberOfReviewers === 0) {
+    return reviewers;
+  }
+
+  return sampleSize(reviewers, numberOfReviewers);
+}
+
 /* Private */
 
 function replace_groups_with_individuals({ reviewers, config }) {
@@ -125,4 +143,5 @@ module.exports = {
   identify_reviewers_by_author,
   should_request_review,
   fetch_default_reviwers,
+  randomly_pick_reviewers,
 };
